Refined definitions of input and output.
Input is what a process consumes.
Output is what a process produces.
A stream is considered a process.
Wait without argument reads the INPUT of the ACTIVE PROCESS.
Send without argument writes the OUTPUT of the ACTIVE PROCESS.
with-io maps the output of the specified source as the active process's input.
with-io maps the input of the specified source as the active process's output.

darcs-hash:20060602215148-68f70-3d9bb14f1716627394216e1298835181e0ee527a